requirements.txt
aiogram==3.4.1
asyncio==3.4.3
python-dotenv==1.0.1
motor==3.4.0
